Binds delete to parent components workExp and Educ
This commits creates the binding for handleDelete
so it works on the parent componentes when invoked from
their respectives childs (work in workExperience &
StudyItem in education).

It took a while to understand everything needed but this is the
resume in case someone is reading this specific commit:
the method in the parents body, the biding as this works with
the parents state, sending as a props, a way to invoke the method
from the child.

Thats done. Going for stilyng now. (Not the best commit).